bdev: forward-port ReAttachNet/DisconnectNet
This is plain copy of the 1.2 ReAttachNet and DisconnectNet methods on
the DRBD8 device, with the logger to logging module changes and the
ReAttachNet method renamed to AttachNet.

These methods are not used anywhere right now, but will be used for
migration and a simpler disk-replace.

The code was originally committed on the 1.2 branch as revision numbers
1165 and 1204.

Originally-Reviewed-by: imsnah, ultrotter