burnin: introduce instance alive checks
This patch adds instance alive checks after most start operations. The
check is done in a custom way:
  - the instance is expected to have an http server up and running
  - and it should server the '/hostname.txt' resource containing the
    hostname of the instance

This allows checking that:
  - creation is working OK
  - start after failover (and in the future migrate) is ok
  - rename works correctly

By default, the check is disabled since one needs a custom OS for this
check.

The patch also fixes a wrong variable name from a previous burnin patch.

Reviewed-by: ultrotter