KVM: add a _CONF_DIR
Currently we keep pid files and control files. In the conf dir we'll
also keep the data to start the instance anew, and the network
interface scripts. These will then be copied to a separate area (since
_CONF_DIR could be mounted 'noexec') and used to start the instance.

This patch also adds comments to state what the various directories are
used for.

Reviewed-by: iustinp