KVM: add the _InstancePidAlive function
Throughout the kvm code we very often look for the instance pidfile
name, read it, and check if the process is alive. Abstract this into a
private function and use that one instead.

This patch also changes RebootInstance to check whether the instance is
alive before trying to reboot it.

Reviewed-by: iustinp